+++ src/Silex/Application.php
@@ -11,6 +11,8 @@
 
 namespace Silex;
 
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpKernel\HttpKernel;
@@ -19,8 +21,6 @@ use Symfony\Component\HttpKernel\TerminableInterface;
 use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\Event\PostResponseEvent;
-use Symfony\Component\HttpKernel\EventListener\ResponseListener;
-use Symfony\Component\HttpKernel\EventListener\RouterListener;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\HttpFoundation\Request;
@@ -30,20 +30,20 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\RouteCollection;
-use Symfony\Component\Routing\RequestContext;
-use Silex\EventListener\LocaleListener;
-use Silex\EventListener\MiddlewareListener;
-use Silex\EventListener\ConverterListener;
-use Silex\EventListener\StringToResponseListener;
+use Silex\Api\BootableProviderInterface;
+use Silex\Api\EventListenerProviderInterface;
+use Silex\Api\ControllerProviderInterface;
+use Silex\Provider\RoutingServiceProvider;
+use Silex\Provider\KernelServiceProvider;
 
 /**
  * The Silex framework class.
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class Application extends \Pimple implements HttpKernelInterface, TerminableInterface
+class Application extends Container implements HttpKernelInterface, TerminableInterface
 {
-    const VERSION = '1.3.2-DEV';
+    const VERSION = '2.0.0-DEV';
 
     const EARLY_EVENT = 512;
     const LATE_EVENT = -512;
@@ -62,96 +62,55 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
     {
         parent::__construct();
 
-        $app = $this;
-
-        $this['logger'] = null;
-
-        $this['routes'] = $this->share(function () {
+        $this['routes_factory'] = $this->factory(function () {
             return new RouteCollection();
         });
 
-        $this['controllers'] = $this->share(function () use ($app) {
-            return $app['controllers_factory'];
-        });
-
-        $this['controllers_factory'] = function () use ($app) {
-            return new ControllerCollection($app['route_factory']);
+        $this['routes'] = function ($app) {
+            return $app['routes_factory'];
         };
+
+        $this['controllers'] = function ($app) {
+            return $app['controllers_factory'];
+        };
+
+        $this['controllers_factory'] = $this->factory(function ($app) {
+            return new ControllerCollection($app['route_factory'], $app['routes_factory']);
+        });
 
         $this['route_class'] = 'Silex\\Route';
-        $this['route_factory'] = function () use ($app) {
+        $this['route_factory'] = $this->factory(function ($app) {
             return new $app['route_class']();
+        });
+
+        $this['exception_handler'] = function ($app) {
+            return new ExceptionHandler($app['debug']);
         };
 
-        $this['exception_handler'] = $this->share(function () use ($app) {
-            return new ExceptionHandler($app['debug']);
-        });
-
-        $this['dispatcher_class'] = 'Symfony\\Component\\EventDispatcher\\EventDispatcher';
-        $this['dispatcher'] = $this->share(function () use ($app) {
-            /*
-             * @var EventDispatcherInterface
-             */
-            $dispatcher = new $app['dispatcher_class']();
-
-            $urlMatcher = new LazyUrlMatcher(function () use ($app) {
-                return $app['url_matcher'];
-            });
-            $dispatcher->addSubscriber(new RouterListener($urlMatcher, $app['request_context'], $app['logger'], $app['request_stack']));
-            $dispatcher->addSubscriber(new LocaleListener($app, $urlMatcher, $app['request_stack']));
-            if (isset($app['exception_handler'])) {
-                $dispatcher->addSubscriber($app['exception_handler']);
-            }
-            $dispatcher->addSubscriber(new ResponseListener($app['charset']));
-            $dispatcher->addSubscriber(new MiddlewareListener($app));
-            $dispatcher->addSubscriber(new ConverterListener($app['routes'], $app['callback_resolver']));
-            $dispatcher->addSubscriber(new StringToResponseListener());
-
-            return $dispatcher;
-        });
-
-        $this['callback_resolver'] = $this->share(function () use ($app) {
+        $this['callback_resolver'] = function ($app) {
             return new CallbackResolver($app);
-        });
+        };
 
-        $this['resolver'] = $this->share(function () use ($app) {
+        $this['resolver'] = function ($app) {
             return new ControllerResolver($app, $app['logger']);
-        });
+        };
 
-        $this['kernel'] = $this->share(function () use ($app) {
+        $this['kernel'] = function ($app) {
             return new HttpKernel($app['dispatcher'], $app['resolver'], $app['request_stack']);
-        });
+        };
 
-        $this['request_stack'] = $this->share(function () use ($app) {
-            if (class_exists('Symfony\Component\HttpFoundation\RequestStack')) {
-                return new RequestStack();
-            }
-        });
-
-        $this['request_context'] = $this->share(function () use ($app) {
-            $context = new RequestContext();
-
-            $context->setHttpPort($app['request.http_port']);
-            $context->setHttpsPort($app['request.https_port']);
-
-            return $context;
-        });
-
-        $this['url_matcher'] = $this->share(function () use ($app) {
-            return new RedirectableUrlMatcher($app['routes'], $app['request_context']);
-        });
-
-        $this['request_error'] = $this->protect(function () {
-            throw new \RuntimeException('Accessed request service outside of request scope. Try moving that call to a before handler or controller.');
-        });
-
-        $this['request'] = $this['request_error'];
+        $this['request_stack'] = function () {
+            return new RequestStack();
+        };
 
         $this['request.http_port'] = 80;
         $this['request.https_port'] = 443;
         $this['debug'] = false;
         $this['charset'] = 'UTF-8';
-        $this['locale'] = 'en';
+        $this['logger'] = null;
+
+        $this->register(new KernelServiceProvider());
+        $this->register(new RoutingServiceProvider());
 
         foreach ($values as $key => $value) {
             $this[$key] = $value;
@@ -170,11 +129,7 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
     {
         $this->providers[] = $provider;
 
-        $provider->register($this);
-
-        foreach ($values as $key => $value) {
-            $this[$key] = $value;
-        }
+        parent::register($provider, $values);
 
         return $this;
     }
@@ -187,12 +142,20 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
      */
     public function boot()
     {
-        if (!$this->booted) {
-            foreach ($this->providers as $provider) {
-                $provider->boot($this);
+        if ($this->booted) {
+            return;
+        }
+
+        $this->booted = true;
+
+        foreach ($this->providers as $provider) {
+            if ($provider instanceof EventListenerProviderInterface) {
+                $provider->subscribe($this, $this['dispatcher']);
             }
 
-            $this->booted = true;
+            if ($provider instanceof BootableProviderInterface) {
+                $provider->boot($this);
+            }
         }
     }
 
@@ -305,11 +268,11 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
             return;
         }
 
-        $this['dispatcher'] = $this->share($this->extend('dispatcher', function (EventDispatcherInterface $dispatcher, $app) use ($callback, $priority, $eventName) {
+        $this->extend('dispatcher', function (EventDispatcherInterface $dispatcher, $app) use ($callback, $priority, $eventName) {
             $dispatcher->addListener($eventName, $app['callback_resolver']->resolveCallback($callback), $priority);
 
             return $dispatcher;
-        }));
+        });
     }
 
     /**
@@ -326,7 +289,7 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
         $app = $this;
 
         $this->on(KernelEvents::REQUEST, function (GetResponseEvent $event) use ($callback, $app) {
-            if (HttpKernelInterface::MASTER_REQUEST !== $event->getRequestType()) {
+            if (!$event->isMasterRequest()) {
                 return;
             }
 
@@ -352,7 +315,7 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
         $app = $this;
 
         $this->on(KernelEvents::RESPONSE, function (FilterResponseEvent $event) use ($callback, $app) {
-            if (HttpKernelInterface::MASTER_REQUEST !== $event->getRequestType()) {
+            if (!$event->isMasterRequest()) {
                 return;
             }
 
@@ -436,12 +399,10 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
 
     /**
      * Flushes the controller collection.
-     *
-     * @param string $prefix The route prefix
      */
-    public function flush($prefix = '')
+    public function flush()
     {
-        $this['routes']->addCollection($this['controllers']->flush($prefix));
+        $this['routes']->addCollection($this['controllers']->flush());
     }
 
     /**
@@ -509,8 +470,6 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
      * @param null|string         $contentDisposition The type of Content-Disposition to set automatically with the filename
      *
      * @return BinaryFileResponse
-     *
-     * @throws \RuntimeException When the feature is not supported, before http-foundation v2.2
      */
     public function sendFile($file, $status = 200, array $headers = array(), $contentDisposition = null)
     {
@@ -524,6 +483,8 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
      * @param ControllerCollection|ControllerProviderInterface $controllers A ControllerCollection or a ControllerProviderInterface instance
      *
      * @return Application
+     *
+     * @throws \LogicException
      */
     public function mount($prefix, $controllers)
     {
@@ -572,17 +533,9 @@ class Application extends \Pimple implements HttpKernelInterface, TerminableInte
             $this->boot();
         }
 
-        $current = HttpKernelInterface::SUB_REQUEST === $type ? $this['request'] : $this['request_error'];
-
-        $this['request'] = $request;
-
         $this->flush();
 
-        $response = $this['kernel']->handle($request, $type, $catch);
-
-        $this['request'] = $current;
-
-        return $response;
+        return $this['kernel']->handle($request, $type, $catch);
     }
 
     /**