+++ src/Silex/ControllerCollection.php
@@ -42,10 +42,12 @@ class ControllerCollection
     protected $defaultRoute;
     protected $defaultController;
     protected $prefix;
+    protected $routesFactory;
 
-    public function __construct(Route $defaultRoute)
+    public function __construct(Route $defaultRoute, $routesFactory = null)
     {
         $this->defaultRoute = $defaultRoute;
+        $this->routesFactory = $routesFactory;
         $this->defaultController = function (Request $request) {
             throw new \LogicException(sprintf('The "%s" route must have code to run when it matches.', $request->attributes->get('_route')));
         };
@@ -171,9 +173,7 @@ class ControllerCollection
         call_user_func_array(array($this->defaultRoute, $method), $arguments);
 
         foreach ($this->controllers as $controller) {
-            if ($controller instanceof Controller) {
-                call_user_func_array(array($controller, $method), $arguments);
-            }
+            call_user_func_array(array($controller, $method), $arguments);
         }
 
         return $this;
@@ -182,13 +182,17 @@ class ControllerCollection
     /**
      * Persists and freezes staged controllers.
      *
-     * @param string $prefix
-     *
      * @return RouteCollection A RouteCollection instance
      */
-    public function flush($prefix = '')
+    public function flush()
     {
-        return $this->doFlush($prefix, new RouteCollection());
+        if (null === $this->routesFactory) {
+            $routes = new RouteCollection();
+        } else {
+            $routes = $this->routesFactory;
+        }
+
+        return $this->doFlush('', $routes);
     }
 
     private function doFlush($prefix, RouteCollection $routes)
@@ -201,9 +205,10 @@ class ControllerCollection
             if ($controller instanceof Controller) {
                 $controller->getRoute()->setPath($prefix.$controller->getRoute()->getPath());
                 if (!$name = $controller->getRouteName()) {
-                    $name = $controller->generateRouteName('');
+                    $name = $base = $controller->generateRouteName('');
+                    $i = 0;
                     while ($routes->get($name)) {
-                        $name .= '_';
+                        $name = $base.'_'.++$i;
                     }
                     $controller->bind($name);
                 }