+++ tests/Silex/Tests/Provider/TwigServiceProviderTest.php
@@ -13,7 +13,6 @@ namespace Silex\Tests\Provider;
 
 use Silex\Application;
 use Silex\Provider\TwigServiceProvider;
-use Silex\Provider\HttpFragmentServiceProvider;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -40,35 +39,6 @@ class TwigServiceProviderTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('Hello john!', $response->getContent());
     }
 
-    public function testRenderFunction()
-    {
-        if (!class_exists('Symfony\Component\HttpFoundation\RequestStack')) {
-            $this->markTestSkipped();
-        }
-
-        $app = new Application();
-
-        $app->register(new HttpFragmentServiceProvider());
-        $app->register(new TwigServiceProvider(), array(
-            'twig.templates' => array(
-                'hello' => '{{ render("/foo") }}',
-                'foo' => 'foo',
-            ),
-        ));
-
-        $app->get('/hello', function () use ($app) {
-            return $app['twig']->render('hello');
-        });
-
-        $app->get('/foo', function () use ($app) {
-            return $app['twig']->render('foo');
-        });
-
-        $request = Request::create('/hello');
-        $response = $app->handle($request);
-        $this->assertEquals('foo', $response->getContent());
-    }
-
     public function testLoaderPriority()
     {
         $app = new Application();
@@ -77,9 +47,9 @@ class TwigServiceProviderTest extends \PHPUnit_Framework_TestCase
         ));
         $loader = $this->getMock('\Twig_LoaderInterface');
         $loader->expects($this->never())->method('getSource');
-        $app['twig.loader.filesystem'] = $app->share(function ($app) use ($loader) {
+        $app['twig.loader.filesystem'] = function ($app) use ($loader) {
             return $loader;
-        });
+        };
         $this->assertEquals('foo', $app['twig.loader']->getSource('foo'));
     }
 }